requirements.txt
# ============================================================
# Emotion Detection 2.0 — Dependencies
# ============================================================
# Install with:  pip install -r requirements.txt
# Designed for CPU-only machines (no GPU required).
# ============================================================

# Core ML & Vision
opencv-python>=4.8.0        # Webcam access, face detection (Haar Cascade), drawing
tensorflow-cpu>=2.15.0      # Keras CNN model — CPU-optimized build (no CUDA needed)
numpy>=1.24.0               # Array operations for image pixel manipulation

# Web Application
fastapi>=0.110.0            # High-performance async web framework
uvicorn[standard]>=0.27.0   # ASGI server for FastAPI
websockets>=12.0            # WebSocket support for real-time streaming
